Refactor memory management and allocator modules
- Updated `Cargo.toml` to reorganize dependencies into mandatory and optional categories, enhancing modularity.
- Modified `build.rs` to conditionally compile user programs based on target architecture (RISC-V and LoongArch).
- Enhanced console output functions to support device features conditionally.
- Cleaned up CPU task management code by removing redundant comments.
- Introduced feature gates for process-related modules in the task management system.
- Updated main module to reflect multi-architecture support and reorganized module imports.
- Moved frame allocator implementation to a new file for better organization and clarity.
- Removed the global allocator module and replaced it with a new `talc_alloc` module for dynamic heap memory allocation.
- Ensured that heap initialization is only called when the allocation feature is enabled.
