+++ os/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-talc = { version = "4" }
+# 基础依赖（始终需要）
 lock_api = "0.4"
 bitflags = { version = "2.10.0" }
 lazy_static = { version = "1.4.0", features = ["spin_no_std"] }
@@ -12,25 +12,30 @@ once_cell = { version = "1.21.3", default-features = false, features = [
     "alloc",
     "race",
 ] }
-xmas-elf = "0.9"
 paste = "1.0"
-ext4_rs = "1.3.2"
-virtio-drivers = "0.12.0"
-fdt = "0.1.5"
-smoltcp = { version = "0.12.0", default-features = false, features = ["alloc", "medium-ethernet", "proto-ipv4", "proto-ipv6", "socket-raw", "socket-tcp", "socket-udp"] }
-chrono = { version = "0.4", default-features = false, features = ["alloc"] }
-uart_16550 = "0.4.0"
-hashbrown = "0.16.1"
+
+# 功能分层依赖（optional，通过 feature gate 引入）
+talc = { version = "4", default-features = false, features = ["nightly_api", "lock_api"], optional = true }
+hashbrown = { version = "0.16.1", optional = true }
+xmas-elf = { version = "0.9", optional = true }
+ext4_rs = { version = "1.3.2", optional = true }
+virtio-drivers = { version = "0.12.0", optional = true }
+fdt = { version = "0.1.5", optional = true }
+smoltcp = { version = "0.12.0", default-features = false, features = ["alloc", "medium-ethernet", "proto-ipv4", "proto-ipv6", "socket-raw", "socket-tcp", "socket-udp"], optional = true }
+chrono = { version = "0.4", default-features = false, features = ["alloc"], optional = true }
+uart_16550 = { version = "0.4.0", optional = true }
 
 [features]
-default = ["sync", "alloc", "paging", "proc_vm", "fs", "net"]
+default = ["sync", "alloc", "paging", "proc_vm", "fs", "net", "device"]
 sync = []
-alloc = ["sync"]
+alloc = ["sync", "dep:talc"]
+# device: 驱动层（virtio, fdt, uart, rtc）
+device = ["alloc", "dep:virtio-drivers", "dep:fdt", "dep:uart_16550", "dep:chrono"]
 paging = ["alloc"]
-proc = []
-fs = ["proc", "sync"]
+proc = ["sync", "dep:hashbrown", "dep:xmas-elf"]
+fs = ["proc", "device", "dep:ext4_rs"]
 proc_vm = ["paging", "fs"]
-net = ["sync"]
+net = ["sync", "device", "dep:smoltcp"]
 proto-ipv6 = []
 
 # RISC-V 架构特定依赖
