+++ os/src/kernel/cpu.rs
@@ -81,9 +81,6 @@ impl Cpu {
         }
 
         // 同步 TrapFrame 的 cpu_ptr 指向当前 CPU，确保多核迁移后 trap_entry 恢复正确的 tp
-        // 说明：trap_entry 会从 TrapFrame.cpu_ptr 恢复 tp，如果任务在不同 CPU 之间迁移，
-        // 需要将该字段更新为当前 CPU 的地址，否则进入内核后会把 tp 设置为错误的 CPU，
-        // 导致 current_cpu()/per-CPU 数据读取混乱乃至崩溃。
         let tf_usize = {
             use core::sync::atomic::Ordering;
             task.lock().trap_frame_ptr.load(Ordering::SeqCst) as usize
